refactor(runtime tests): split capture test suites
